TEST Unit test for some basic functions
